servoshell

Module desktop

Contains files specific to the servoshell app for Desktop systems.

Modules

accelerated_gl_media 🔒

app 🔒

Application entry point, runs the event loop.

cli 🔒

dialog 🔒

event_loop 🔒

An event loop implementation that works in headless mode.

gamepad 🔒

geometry

gui 🔒

headed_window 🔒

A winit window implementation.

headless_window 🔒

A headless window implementation.

keyutils 🔒

protocols 🔒

tracing 🔒

webxr 🔒