servoshell::desktop::gui

Struct Gui

pub struct Gui {
    rendering_context: Rc<OffscreenRenderingContext>,
    context: EguiGlow,
    toolbar_height: Length<f32, DeviceIndependentPixel>,
    location: String,
    location_dirty: bool,
    load_status: LoadStatus,
    status_text: Option<String>,
    can_go_back: bool,
    can_go_forward: bool,
    favicon_textures: HashMap<WebViewId, (TextureHandle, SizedTexture)>,
    pending_accesskit_updates: Vec<TreeUpdate>,
}

The user interface of a headed servoshell. Currently this is implemented via egui.

Fields
rendering_context: Rc<OffscreenRenderingContext>context: EguiGlowtoolbar_height: Length<f32, DeviceIndependentPixel>location: Stringlocation_dirty: bool

Whether the location has been edited by the user without clicking Go.

load_status: LoadStatus

The LoadStatus of the active WebView.

status_text: Option<String>

The text to display in the status bar on the bottom of the window.

can_go_back: bool

Whether or not the current WebView can navigate backward.

can_go_forward: bool

Whether or not the current WebView can navigate forward.

favicon_textures: HashMap<WebViewId, (TextureHandle, SizedTexture)>

Handle to the GPU texture of the favicon.

These need to be cached across egui draw calls.

pending_accesskit_updates: Vec<TreeUpdate>

AccessKit tree updates pending the next egui tick. This allows us to ensure that graft nodes are sent before the subtrees they graft.

Implementations

impl Gui

pub(crate) fn new( winit_window: &Window, event_loop: &ActiveEventLoop, event_loop_proxy: EventLoopProxy<AppEvent>, rendering_context: Rc<OffscreenRenderingContext>, initial_url: Url, ) -> Self

pub(crate) fn has_keyboard_focus(&self) -> bool

pub(crate) fn surrender_focus(&self)

pub(crate) fn on_window_event( &mut self, winit_window: &Window, event: &WindowEvent, ) -> EventResponse

pub(crate) fn toolbar_height(&self) -> Length<f32, DeviceIndependentPixel>

The height of the top toolbar of this user inteface ie the distance from the top of the window to the position of the WebView.

pub(crate) fn is_in_egui_toolbar_rect( &self, position: Point2D<f32, DeviceIndependentPixel>, ) -> bool

Return true iff the given position is over the egui toolbar.

fn toolbar_button(text: &str) -> Button<'_>

Create a frameless button with square sizing, as used in the toolbar.

fn browser_tab( ui: &mut Ui, window: &ServoShellWindow, webview: WebView, favicon_texture: Option<SizedTexture>, )

Draws a browser tab, checking for clicks and queues appropriate UserInterfaceCommands. Using a custom widget here would’ve been nice, but it doesn’t seem as though egui supports that, so we arrange multiple Widgets in a way that they look connected.

pub(crate) fn update( &mut self, state: &RunningAppState, window: &ServoShellWindow, headed_window: &HeadedWindow, )

Update the user interface, but do not paint the updated state.

pub(crate) fn paint(&mut self, window: &Window)

Paint the GUI, as of the last update.

fn update_location_in_toolbar(&mut self, window: &ServoShellWindow) -> bool

Updates the location field from the given RunningAppState, unless the user has started editing it without clicking Go, returning true iff it has changed (needing an egui update).

fn update_load_status(&mut self, window: &ServoShellWindow) -> bool

fn update_status_text(&mut self, window: &ServoShellWindow) -> bool

fn update_can_go_back_and_forward(&mut self, window: &ServoShellWindow) -> bool

pub(crate) fn update_webview_data(&mut self, window: &ServoShellWindow) -> bool

Updates all fields taken from the given ServoShellWindow, such as the location field. Returns true iff the egui needs an update.

pub(crate) fn handle_accesskit_event(&mut self, event: &WindowEvent) -> bool

Returns true if a redraw is required after handling the provided event.

pub(crate) fn set_zoom_factor(&self, factor: f32)

pub(crate) fn notify_accessibility_tree_update( &mut self, tree_update: TreeUpdate, )