servoshell::desktop::headed_window

Struct HeadedWindow

pub struct HeadedWindow {

    gui: RefCell<Gui>,
    screen_size: Size2D<u32, DeviceIndependentPixel>,
    webview_relative_mouse_point: Cell<Point2D<f32, DevicePixel>>,
    inner_size: Cell<PhysicalSize<u32>>,
    fullscreen: Cell<bool>,
    fullscreen_from_document: Cell<bool>,
    device_pixel_ratio_override: Option<f32>,
    xr_window_poses: RefCell<Vec<Rc<XRWindowPose>>>,
    modifiers_state: Cell<ModifiersState>,
    rendering_context: Rc<OffscreenRenderingContext>,
    window_rendering_context: Rc<WindowRenderingContext>,
    touch_event_simulator: Option<TouchEventSimulator>,
    pending_keyboard_events: RefCell<HashMap<InputEventId, KeyboardEvent>>,
    winit_window: Window,
    last_title: RefCell<String>,
    dialogs: RefCell<HashMap<WebViewId, Vec<Dialog>>>,
    visible_input_method: Cell<Option<EmbedderControlId>>,
    last_mouse_position: Cell<Option<Point2D<f32, DeviceIndependentPixel>>>,

}

Fields
gui: RefCell<Gui>

The egui interface that is responsible for showing the user interface elements of this headed Window.

screen_size: Size2D<u32, DeviceIndependentPixel>webview_relative_mouse_point: Cell<Point2D<f32, DevicePixel>>inner_size: Cell<PhysicalSize<u32>>

The inner size of the window in physical pixels which excludes OS decorations. It equals viewport size + (0, toolbar height).

fullscreen: Cell<bool>fullscreen_from_document: Cell<bool>device_pixel_ratio_override: Option<f32>xr_window_poses: RefCell<Vec<Rc<XRWindowPose>>>modifiers_state: Cell<ModifiersState>rendering_context: Rc<OffscreenRenderingContext>

The RenderingContext of Servo itself. This is used to render Servo results temporarily until they can be blitted into the egui scene.

window_rendering_context: Rc<WindowRenderingContext>

The RenderingContext that renders directly onto the Window. This is used as the target of egui rendering and also where Servo rendering results are finally blitted.

touch_event_simulator: Option<TouchEventSimulator>

A helper that simulates touch events when the --simulate-touch-events flag is enabled.

pending_keyboard_events: RefCell<HashMap<InputEventId, KeyboardEvent>>

Keyboard events that have been sent to Servo that have still not been handled yet. When these are handled, they will optionally be used to trigger keybindings that are overridable by web content.

winit_window: Windowlast_title: RefCell<String>

The last title set on this window. We need to store this value here, as winit::Window::title is not supported very many platforms.

dialogs: RefCell<HashMap<WebViewId, Vec<Dialog>>>

The current set of open dialogs.

visible_input_method: Cell<Option<EmbedderControlId>>

The EmbedderControlId of the currently showing [InputMethod] interfaces, if one is showing.

last_mouse_position: Cell<Option<Point2D<f32, DeviceIndependentPixel>>>

The position of the mouse cursor after the most recent MouseMove event.

Implementations

impl HeadedWindow

pub(crate) fn new( servoshell_preferences: &ServoShellPreferences, event_loop: &ActiveEventLoop, event_loop_proxy: EventLoopProxy<AppEvent>, initial_url: Url, ) -> Rc<Self>

pub(crate) fn winit_window(&self) -> &Window

fn handle_keyboard_input( &self, state: Rc<RunningAppState>, window: &Rc<ServoShellWindow>, winit_event: KeyEvent, )

fn handle_mouse_button_event( &self, webview: &WebView, button: MouseButton, action: ElementState, )

Helper function to handle a click

fn handle_mouse_move_event( &self, webview: &WebView, position: PhysicalPosition<f64>, )

Helper function to handle mouse move events.

fn handle_intercepted_key_bindings( &self, state: Rc<RunningAppState>, window: &Rc<ServoShellWindow>, key_event: &KeyboardEvent, ) -> bool

Handle key events before sending them to Servo.

fn force_srgb_color_space(window_handle: RawWindowHandle)

fn show_ime( &self, control_id: EmbedderControlId, input_method: InputMethodControl, )

pub(crate) fn for_each_active_dialog( &self, window: &ServoShellWindow, callback: impl Fn(&mut Dialog) -> bool, )

fn add_dialog(&self, webview_id: WebViewId, dialog: Dialog)

fn remove_dialog( &self, webview_id: WebViewId, embedder_control_id: EmbedderControlId, )

fn has_active_dialog_for_webview(&self, webview_id: WebViewId) -> bool

fn toolbar_height(&self) -> Length<f32, DeviceIndependentPixel>

pub(crate) fn handle_winit_window_event( &self, state: Rc<RunningAppState>, window: Rc<ServoShellWindow>, event: WindowEvent, )

pub(crate) fn handle_winit_app_event( &self, state: Rc<RunningAppState>, app_event: AppEvent, )

pub(crate) fn is_fullscreen_from_document(&self) -> bool