Struct servoshell::desktop::minibrowser::Minibrowser

pub struct Minibrowser {
    pub context: EguiGlow,
    pub event_queue: RefCell<Vec<MinibrowserEvent>>,
    pub toolbar_height: Length<f32, DeviceIndependentPixel>,
    widget_surface_fbo: Option<NativeFramebuffer>,
    last_update: Instant,
    last_mouse_position: Option<Point2D<f32, DeviceIndependentPixel>>,
    location: RefCell<String>,
    location_dirty: Cell<bool>,
    load_status: LoadStatus,
    status_text: Option<String>,
}

Fields
context: EguiGlowevent_queue: RefCell<Vec<MinibrowserEvent>>toolbar_height: Length<f32, DeviceIndependentPixel>widget_surface_fbo: Option<NativeFramebuffer>

The framebuffer object name for the widget surface we should draw to, or None if our widget surface does not use a framebuffer object.

last_update: Instantlast_mouse_position: Option<Point2D<f32, DeviceIndependentPixel>>location: RefCell<String>location_dirty: Cell<bool>

Whether the location has been edited by the user without clicking Go.

load_status: LoadStatusstatus_text: Option<String>
Implementations

impl Minibrowser

pub fn new( rendering_context: &RenderingContext, event_loop: &EventLoop<WakerEvent>, initial_url: ServoUrl, ) -> Self

pub fn on_window_event( &mut self, window: &Window, event: &WindowEvent, ) -> EventResponse

Preprocess the given winit::event::WindowEvent, returning unconsumed for mouse events in the Servo browser rect. This is needed because the CentralPanel we create for our webview would otherwise make egui report events in that area as consumed.

fn is_in_browser_rect( &self, position: Point2D<f32, DeviceIndependentPixel>, ) -> bool

Return true iff the given position is in the Servo browser rect.

fn toolbar_button(text: &str) -> Button<'_>

Create a frameless button with square sizing, as used in the toolbar.

fn browser_tab( ui: &mut Ui, label: &str, selected: bool, webview_id: TopLevelBrowsingContextId, ) -> Option<EmbedderEvent>

Draws a browser tab, checking for clicks and returns an appropriate EmbedderEvent Using a custom widget here would’ve been nice, but it doesn’t seem as though egui supports that, so we arrange multiple Widgets in a way that they look connected.

pub fn update( &mut self, window: &Window, webviews: &mut WebViewManager<dyn WindowPortsMethods>, servo_framebuffer_id: Option<GLuint>, reason: &'static str, )

Update the minibrowser, but don’t paint. If servo_framebuffer_id is given, set up a paint callback to blit its contents to our CentralPanel when Minibrowser::paint is called.

pub fn paint(&mut self, window: &Window)

Paint the minibrowser, as of the last update.

pub fn queue_embedder_events_for_minibrowser_events( &self, browser: &WebViewManager<dyn WindowPortsMethods>, app_event_queue: &mut Vec<EmbedderEvent>, )

Takes any outstanding events from the Minibrowser, converting them to EmbedderEvent and routing those to the App event queue.

pub fn update_location_in_toolbar( &mut self, browser: &mut WebViewManager<dyn WindowPortsMethods>, ) -> bool

Updates the location field from the given WebViewManager, unless the user has started editing it without clicking Go, returning true iff it has changed (needing an egui update).

pub fn update_spinner_in_toolbar( &mut self, browser: &mut WebViewManager<dyn WindowPortsMethods>, ) -> bool

Updates the spinner from the given WebViewManager, returning true iff it has changed (needing an egui update).

pub fn update_status_text( &mut self, browser: &mut WebViewManager<dyn WindowPortsMethods>, ) -> bool

pub fn update_webview_data( &mut self, browser: &mut WebViewManager<dyn WindowPortsMethods>, ) -> bool

Updates all fields taken from the given WebViewManager, such as the location field. Returns true iff the egui needs an update.