Struct servoshell::desktop::webview::WebViewManager

pub struct WebViewManager<Window: WindowPortsMethods + ?Sized> {
    status_text: Option<String>,
    webviews: HashMap<TopLevelBrowsingContextId, WebView>,
    creation_order: Vec<TopLevelBrowsingContextId>,
    focused_webview_id: Option<TopLevelBrowsingContextId>,
    webview_preload_data: HashMap<TopLevelBrowsingContextId, WebViewPreloadData>,
    window: Rc<Window>,
    event_queue: Vec<EmbedderEvent>,
    clipboard: Option<Clipboard>,
    gamepad: Option<Gilrs>,
    haptic_effects: HashMap<usize, HapticEffect>,
    shutdown_requested: bool,
}

Fields
status_text: Option<String>webviews: HashMap<TopLevelBrowsingContextId, WebView>

List of top-level browsing contexts. Modified by EmbedderMsg::WebViewOpened and EmbedderMsg::WebViewClosed, and we exit if it ever becomes empty.

creation_order: Vec<TopLevelBrowsingContextId>

The order in which the webviews were created.

focused_webview_id: Option<TopLevelBrowsingContextId>

The webview that is currently focused. Modified by EmbedderMsg::WebViewFocused and EmbedderMsg::WebViewBlurred.

webview_preload_data: HashMap<TopLevelBrowsingContextId, WebViewPreloadData>

Pre-creation state for WebViews. This is needed because in some situations the WebViewOpened event is sent after ChangePageTitle and HistoryChanged

window: Rc<Window>event_queue: Vec<EmbedderEvent>clipboard: Option<Clipboard>gamepad: Option<Gilrs>haptic_effects: HashMap<usize, HapticEffect>shutdown_requested: bool
Implementations

impl<Window> WebViewManager<Window>
where Window: WindowPortsMethods + ?Sized,

pub fn new(window: Rc<Window>) -> WebViewManager<Window>

pub fn get_mut(&mut self, webview_id: WebViewId) -> Option<&mut WebView>

fn ensure_preload_data_mut( &mut self, webview_id: &WebViewId, ) -> &mut WebViewPreloadData

pub fn focused_webview_id(&self) -> Option<WebViewId>

pub fn current_url_string(&self) -> Option<String>

pub fn focused_webview(&self) -> Option<&WebView>

pub fn load_status(&self) -> LoadStatus

pub fn status_text(&self) -> Option<String>

pub fn get_events(&mut self) -> Vec<EmbedderEvent>

pub fn webviews(&self) -> Vec<(WebViewId, &WebView)>

pub fn handle_window_events(&mut self, events: Vec<EmbedderEvent>)

pub fn handle_gamepad_events(&mut self)

Handle updates to connected gamepads from GilRs

fn map_gamepad_button(button: Button) -> usize

fn play_haptic_effect( &mut self, index: usize, params: DualRumbleEffectParams, effect_complete_sender: IpcSender<bool>, )

fn stop_haptic_effect(&mut self, index: usize) -> bool

pub fn shutdown_requested(&self) -> bool

fn focus_webview_by_index(&self, index: usize) -> Option<EmbedderEvent>

fn get_focused_webview_index(&self) -> Option<usize>

fn handle_key_from_window(&mut self, key_event: KeyboardEvent)

Handle key events before sending them to Servo.

fn handle_key_from_servo( &mut self, webview_id: Option<WebViewId>, event: KeyboardEvent, )

Handle key events after they have been handled by Servo.

fn scroll_window_from_key( &mut self, scroll_location: ScrollLocation, phase: TouchEventType, webview_id: Option<WebViewId>, )

pub fn handle_servo_events( &mut self, events: Drain<'_, (Option<WebViewId>, EmbedderMsg)>, ) -> ServoEventResponse

Returns true if the caller needs to manually present a new frame.