servoshell::parser

Function location_bar_input_to_url

pub(crate) fn location_bar_input_to_url(
    request: &str,
    searchpage: &str,
) -> Option<ServoUrl>

Interpret an input URL.

If this is not a valid URL, try to “fix” it by adding a scheme or if all else fails, interpret the string as a search term.