servoshell

Module prefs

Structs

CmdArgs 🔒

ServoShellPreferences 🔒

Enums

ArgumentParsingResult 🔒

ParseResolutionError 🔒

Statics

EXPERIMENTAL_PREFS 🔒

Preferences enabled when servoshell is launched with the --enable-experimental-web-platform-features flag.

Functions

default_config_dir

flag_with_default_parser 🔒

This is a helper function that fulfills the following parsing task check for long/short cmd. If there is the flag with this If the flag is not there, parse None`` If the flag is there but no argument, parse Some(default)` If the flag is there and an argument parse the argument

get_preferences 🔒

Get a Servo Preferences to use when initializing Servo by first reading the user preferences file and then overriding these preferences with the ones from the --prefs-file command-line argument, if given.

map_debug_options 🔒

parse_arguments_helper 🔒

parse_command_line_arguments 🔒

Parse Commandline arguments

parse_diagnostics_logging 🔒

Parse the ‘-Z’ command-line flags.

parse_resolution_string 🔒

Parse a resolution string into a Size2D.

parse_user_stylesheets 🔒

Parse a space or comma-separated list of stylesheet paths into a vector of UserStyleSheet.

profile 🔒

read_prefs_file 🔒

read_prefs_map

update_preferences_from_command_line_arguments 🔒

userscripts 🔒

webdriver_port 🔒