Struct servoshell::webview::WebViewManager

pub struct WebViewManager<Window: WindowPortsMethods + ?Sized> {
    current_url: Option<ServoUrl>,
    current_url_string: Option<String>,
    webviews: HashMap<TopLevelBrowsingContextId, WebView>,
    creation_order: Vec<TopLevelBrowsingContextId>,
    focused_webview_id: Option<TopLevelBrowsingContextId>,
    title: Option<String>,
    window: Rc<Window>,
    event_queue: Vec<EmbedderEvent>,
    clipboard: Option<Clipboard>,
    gamepad: Option<Gilrs>,
    shutdown_requested: bool,
    load_status: LoadStatus,
}

Fields
current_url: Option<ServoUrl>current_url_string: Option<String>webviews: HashMap<TopLevelBrowsingContextId, WebView>

List of top-level browsing contexts. Modified by EmbedderMsg::WebViewOpened and EmbedderMsg::WebViewClosed, and we exit if it ever becomes empty.

creation_order: Vec<TopLevelBrowsingContextId>

The order in which the webviews were created.

focused_webview_id: Option<TopLevelBrowsingContextId>

The webview that is currently focused. Modified by EmbedderMsg::WebViewFocused and EmbedderMsg::WebViewBlurred.

title: Option<String>window: Rc<Window>event_queue: Vec<EmbedderEvent>clipboard: Option<Clipboard>gamepad: Option<Gilrs>shutdown_requested: boolload_status: LoadStatus
Implementations

impl<Window> WebViewManager<Window>where Window: WindowPortsMethods + ?Sized,

pub fn new(window: Rc<Window>) -> WebViewManager<Window>

pub fn webview_id(&self) -> Option<WebViewId>

pub fn current_url_string(&self) -> Option<&str>

pub fn load_status(&self) -> LoadStatus

pub fn get_events(&mut self) -> Vec<EmbedderEvent>

pub fn handle_window_events(&mut self, events: Vec<EmbedderEvent>)

pub fn handle_gamepad_events(&mut self)

Handle updates to connected gamepads from GilRs

fn map_gamepad_button(button: Button) -> usize

pub fn shutdown_requested(&self) -> bool

fn handle_key_from_window(&mut self, key_event: KeyboardEvent)

Handle key events before sending them to Servo.

fn platform_handle_key(&mut self, key_event: KeyboardEvent)

fn handle_key_from_servo(&mut self, _: Option<WebViewId>, event: KeyboardEvent)

Handle key events after they have been handled by Servo.

fn scroll_window_from_key( &mut self, scroll_location: ScrollLocation, phase: TouchEventType )

pub fn handle_servo_events( &mut self, events: Drain<'_, (Option<WebViewId>, EmbedderMsg)> ) -> ServoEventResponse

Returns true if the caller needs to manually present a new frame.