servoshell

Module window

Structs

ServoShellWindow 🔒

ServoShellWindowId 🔒

Constants

LINE_HEIGHT 🔒

LINE_WIDTH 🔒

MIN_WINDOW_INNER_SIZE 🔒

https://github.com/web-platform-tests/wpt/blob/9320b1f724632c52929a3fdb11bdaf65eafc7611/webdriver/tests/classic/set_window_rect/set.py#L287-L290 “A window size of 10x10px shouldn’t be supported by any browser.”

Statics

SERVOSHELL_WINDOW_ID 🔒

Traits

PlatformWindow 🔒

A PlatformWindow abstracts away the differents kinds of platform windows that might be used in a servoshell execution. This currently includes headed (winit) and headless windows.