servoshell

Module desktop

Contains files specific to the servoshell app for Desktop systems.

Modules

accelerated_gl_media ğŸ”’

app ğŸ”’

Application entry point, runs the event loop.

cli ğŸ”’

dialog ğŸ”’

event_loop ğŸ”’

An event loop implementation that works in headless mode.

gamepad ğŸ”’

geometry

gui ğŸ”’

headed_window ğŸ”’

A winit window implementation.

headless_window ğŸ”’

A headless window implementation.

keyutils ğŸ”’

protocols ğŸ”’

tracing ğŸ”’

webxr ğŸ”’