Module servoshell::desktop

Contains files specific to the servoshell app for Desktop systems.

Modules

app 🔒

Application entry point, runs the event loop.

cli 🔒

egui_glue 🔒

A modified version of EguiGlow from egui_glow 0.22.0 that retains its shapes, allowing EguiGlow::paint to be called multiple times.

embedder 🔒

Implements the global methods required by Servo (not window/gl/compositor related).

events_loop 🔒

An event loop implementation that works in headless mode.

geometry

headed_window 🔒

A winit window implementation.

headless_window 🔒

A headless window implementation.

keyutils 🔒

minibrowser 🔒

protocols 🔒

tracing 🔒

webview 🔒

window_trait 🔒

Definition of Window. Implemented by headless and headed windows.