servoshell

Module prefs

Structs

CmdArgs ğŸ”’

ServoShellPreferences ğŸ”’

Enums

ArgumentParsingResult ğŸ”’

ParseResolutionError ğŸ”’

Statics

EXPERIMENTAL_PREFS ğŸ”’

Functions

cmd_args ğŸ”’

default_config_dir

flag_with_default_parser ğŸ”’

This is a helper function that fulfills the following parsing task check for long/short cmd. If there is the flag with this If the flag is not there, parse None`` If the flag is there but no argument, parse Some(default)` If the flag is there and an argument parse the arugment

get_preferences ğŸ”’

Get a Servo Preferences to use when initializing Servo by first reading the user preferences file and then overriding these preferences with the ones from the --prefs-file command-line argument, if given.

map_debug_options ğŸ”’

parse_arguments_helper ğŸ”’

parse_command_line_arguments ğŸ”’

Parse Commandline arguments

parse_resolution_string ğŸ”’

Parse a resolution string into a Size2D.

parse_user_stylesheets ğŸ”’

Parse stylesheets into the byte stream.

profile ğŸ”’

read_prefs_file ğŸ”’

read_prefs_map

update_preferences_from_command_line_arguemnts ğŸ”’

userscripts ğŸ”’

webdriver_port ğŸ”’