servoshell::running_app_state

Struct RunningAppState

pub(crate) struct RunningAppState {

    gamepad_support: RefCell<Option<GamepadSupport>>,
    pub(crate) webdriver_senders: RefCell<WebDriverSenders>,
    pub(crate) webdriver_embedder_controls: WebDriverEmbedderControls,
    pub(crate) pending_webdriver_events: RefCell<HashMap<InputEventId, Sender<()>>>,
    pub(crate) webdriver_receiver: Option<Receiver<WebDriverCommandMsg>>,
    pub(crate) servoshell_preferences: ServoShellPreferences,
    pub(crate) servo: Servo,
    pub(crate) achieved_stable_image: Rc<Cell<bool>>,
    pub(crate) user_content_manager: Rc<UserContentManager>,
    exit_scheduled: Cell<bool>,
    experimental_preferences_enabled: Cell<bool>,
    windows: RefCell<HashMap<ServoShellWindowId, Rc<ServoShellWindow>>>,
    focused_window: RefCell<Option<Rc<ServoShellWindow>>>,

}

Fields
gamepad_support: RefCell<Option<GamepadSupport>>

Gamepad support, which may be None if it failed to initialize.

webdriver_senders: RefCell<WebDriverSenders>

The WebDriverSenders used to reply to pending WebDriver requests.

webdriver_embedder_controls: WebDriverEmbedderControls

When running in WebDriver mode, WebDriverEmbedderControls is a virtual container for all embedder controls. This overrides the normal behavior where these controls are shown in the GUI or not processed at all in headless mode.

pending_webdriver_events: RefCell<HashMap<InputEventId, Sender<()>>>

A HashMap of pending WebDriver events. It is the WebDriver embedder’s responsibility to inform the WebDriver server when the event has been fully handled. This map is used to report back to WebDriver when that happens.

webdriver_receiver: Option<Receiver<WebDriverCommandMsg>>

A Receiver for receiving commands from a running WebDriver server, if WebDriver was enabled.

servoshell_preferences: ServoShellPreferences

servoshell specific preferences created during startup of the application.

servo: Servo

A handle to the Servo instance.

achieved_stable_image: Rc<Cell<bool>>

Whether or not the application has achieved stable image output. This is used for the exit_after_stable_image option.

user_content_manager: Rc<UserContentManager>

The UserContentManager for all WebViews created.

exit_scheduled: Cell<bool>

Whether or not program exit has been triggered. This means that all windows will be destroyed and shutdown will start at the end of the current event loop.

experimental_preferences_enabled: Cell<bool>

Whether the user has enabled experimental preferences.

windows: RefCell<HashMap<ServoShellWindowId, Rc<ServoShellWindow>>>

The set of ServoShellWindows that currently exist for this instance of servoshell.

focused_window: RefCell<Option<Rc<ServoShellWindow>>>

The currently focused ServoShellWindow, if one is focused.

Implementations

impl RunningAppState

pub(crate) fn new( servo: Servo, servoshell_preferences: ServoShellPreferences, event_loop_waker: Box<dyn EventLoopWaker>, user_content_manager: Rc<UserContentManager>, ) -> Self

pub(crate) fn open_window( self: &Rc<Self>, platform_window: Rc<dyn PlatformWindow>, initial_url: Url, ) -> Rc<ServoShellWindow>

pub(crate) fn windows<'a>( &'a self, ) -> Ref<'a, HashMap<ServoShellWindowId, Rc<ServoShellWindow>>>

pub(crate) fn focused_window(&self) -> Option<Rc<ServoShellWindow>>

pub(crate) fn focus_window(&self, window: Rc<ServoShellWindow>)

pub(crate) fn window( &self, id: ServoShellWindowId, ) -> Option<Rc<ServoShellWindow>>

pub(crate) fn webview_by_id(&self, webview_id: WebViewId) -> Option<WebView>

pub(crate) fn webdriver_receiver( &self, ) -> Option<&Receiver<WebDriverCommandMsg>>

pub(crate) fn servo(&self) -> &Servo

pub(crate) fn schedule_exit(&self)

pub(crate) fn experimental_preferences_enabled(&self) -> bool

pub(crate) fn set_experimental_preferences_enabled(&self, new_value: bool)

fn close_empty_windows(&self)

Close any ServoShellWindow that doesn’t have an open WebView.

pub(crate) fn spin_event_loop( self: &Rc<Self>, create_platform_window: Option<&dyn Fn(Url) -> Rc<dyn PlatformWindow>>, ) -> bool

Spins the internal application event loop.

Notifies Servo about incoming gamepad events

Spin the Servo event loop, which will update Servo’s embedding layer and trigger delegate methods.

Returns true if the event loop should continue spinning and false if it should exit.

pub(crate) fn maybe_window_for_webview_id( &self, webview_id: WebViewId, ) -> Option<Rc<ServoShellWindow>>

pub(crate) fn window_for_webview_id( &self, webview_id: WebViewId, ) -> Rc<ServoShellWindow>

pub(crate) fn platform_window_for_webview_id( &self, webview_id: WebViewId, ) -> Rc<dyn PlatformWindow>

fn maybe_request_screenshot(&self, webview: WebView)

If we are exiting after achieving a stable image or we want to save the display of the WebView to an image file, request a screenshot of the WebView.

pub(crate) fn set_pending_traversal( &self, traversal_id: TraversalId, sender: GenericSender<WebDriverLoadStatus>, )

pub(crate) fn set_load_status_sender( &self, webview_id: WebViewId, sender: GenericSender<WebDriverLoadStatus>, )

fn remove_load_status_sender(&self, webview_id: WebViewId)

fn set_script_command_interrupt_sender( &self, sender: Option<GenericSender<WebDriverJSResult>>, )

pub(crate) fn handle_webdriver_input_event( &self, webview_id: WebViewId, input_event: InputEvent, response_sender: Option<Sender<()>>, )

pub(crate) fn handle_webdriver_screenshot( &self, webview_id: WebViewId, rect: Option<Rect<f32, CSSPixel>>, result_sender: Sender<Result<RgbaImage, ScreenshotCaptureError>>, )

pub(crate) fn handle_webdriver_script_command( &self, script_command: &WebDriverScriptCommand, )

pub(crate) fn handle_webdriver_load_url( &self, webview_id: WebViewId, url: Url, load_status_sender: GenericSender<WebDriverLoadStatus>, )

pub(crate) fn handle_gamepad_events(&self)

pub(crate) fn handle_focused(&self, window: Rc<ServoShellWindow>)

fn interrupt_webdriver_script_evaluation(&self)

Interrupt any ongoing WebDriver-based script evaluation.

From https://w3c.github.io/webdriver/#dfn-execute-a-function-body:

The rules to execute a function body are as follows. The algorithm returns an ECMAScript completion record.

If at any point during the algorithm a user prompt appears, immediately return Completion { Type: normal, Value: null, Target: empty }, but continue to run the other steps of this algorithm in parallel.

impl RunningAppState

pub(crate) fn handle_webdriver_messages( self: &Rc<Self>, create_platform_window: Option<&dyn Fn(Url) -> Rc<dyn PlatformWindow>>, )